refactor: seperate display functions from controller
