+++ README.md
@@ -43,7 +43,8 @@
 - localStorage 적용하기
 - index.js 에서 탭 버튼 화면 분리하기 ✔️
 - machine.js 에서 validate (에러 검증) 함수들 분리하기 ✔️
-- machine.js 에서 display(화면 보여주는 기능) 함수들 분리하기
+- machine.js 에서 display(화면 보여주는 기능) 함수들 분리하기 ✔️
+- 처음 앱이 시작할 때, localStorage를 보고 display 하기
 - displayProductPurchaseTab 이벤트 리스너 분리하기
